requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
configparser==5.0.0
idna==2.8
notify-run==0.0.13
PyQRCode==1.2.1
python-dotenv==0.13.0
requests==2.22.0
soupsieve==2.0.1
urllib3==1.25.9
